Removed duplicate UserPref class
