--- src/main/java/NASA/model/ReadOnlyUserPrefsNasa.java
@@ -1,16 +1,0 @@
-package NASA.model;
-
-import java.nio.file.Path;
-
-import NASA.commons.core.GuiSettings;
-
-/**
- * Unmodifiable view of user prefs.
- */
-public interface ReadOnlyUserPrefsNasa {
-
-    GuiSettings getGuiSettings();
-
-    Path getNasaBookFilePath();
-
-}
